refactor(todo): share agent normalization
